v5.x-scala-js: move SigmaSerializerExceptions to `interpreter` module
