+++ sc/src/main/scala/sigmastate/serialization/TypeSerializer.scala
@@ -3,10 +3,10 @@ package sigmastate.serialization
 import java.nio.charset.StandardCharsets
 import org.ergoplatform.validation.ValidationRules.{CheckPrimitiveTypeCode, CheckTypeCode}
 import sigmastate._
-import sigmastate.lang.exceptions.InvalidTypePrefix
 import sigmastate.util.safeNewArray
 import sigmastate.utils.{SigmaByteReader, SigmaByteWriter}
 import debox.cfor
+import sigmastate.exceptions.InvalidTypePrefix
 
 /** Serialization of types according to specification in TypeSerialization.md. */
 object TypeSerializer {
