+++ sc/src/test/scala/sigmastate/serialization/ErgoTreeSerializerSpecification.scala
@@ -7,9 +7,8 @@ import org.ergoplatform.validation.ValidationRules.CheckDeserializedScriptIsSigm
 import sigmastate.Values.{BigIntConstant, ByteConstant, ConstantPlaceholder, ErgoTree, IntConstant, ShortConstant, SigmaPropValue, UnparsedErgoTree}
 import sigmastate._
 import sigmastate.eval.{CBigInt, IRContext}
-import sigmastate.exceptions.SerializerException
+import sigmastate.exceptions.{SerializerException, ReaderPositionLimitExceeded}
 import sigmastate.helpers.SigmaTestingCommons
-import sigmastate.lang.exceptions.ReaderPositionLimitExceeded
 import sigmastate.serialization.ErgoTreeSerializer.DefaultSerializer
 import sigmastate.utxo.{DeserializeContext, DeserializeRegister}
